更新

+++ src/lzf/webserver/mapper/GlobelMapper.java
@@ -3,21 +3,31 @@ package lzf.webserver.mapper;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import lzf.webserver.Container;
 import lzf.webserver.Context;
 import lzf.webserver.Host;
 import lzf.webserver.Service;
+import lzf.webserver.log.Log;
+import lzf.webserver.log.LogFactory;
 
 /**
 * @author 李子帆
 * @version 1.0
 * @date 2018年7月14日 下午5:02:43
 * @Description 全局URL路由器，每个Service对象拥有一个GlobelMapper对象
+* 以下添加容器的方法中必须通过Mapped监听器完成
 */
 public final class GlobelMapper {
+	
+	private static final Log log = LogFactory.getLog(GlobelMapper.class);
 
 	private final Service service;
 	
+	//主机名和主机Mappr对象的Map
 	private final Map<String, MappedHost> mapper = new LinkedHashMap<>();
+	
+	//主机Mapper和Context容器名和Context的Mapper对象的Map
+	private final Map<MappedHost, Map<String, MappedContext>> hostMapper = new LinkedHashMap<>();
 	
 	public GlobelMapper(Service service) {
 		this.service = service;
@@ -27,24 +37,39 @@ public final class GlobelMapper {
 	 * 添加主机
 	 * @param host
 	 */
-	public synchronized void addHost(Host host) {
+	synchronized void addHost(Host host) {
 		MappedHost mappedHost = new MappedHost(host.getName(), host);
 		mapper.put(host.getName(), mappedHost);
+		
+		Map<String, MappedContext> map = new LinkedHashMap<>();
+		
+		for(Container c : host.getChildContainers()) {
+			MappedContext mc = new MappedContext(c.getName(), (Context)c);
+			map.put(c.getName(), mc);
+		}
+		
+		hostMapper.put(mappedHost, map);
 	}
 	
 	/**
 	 * 根据Host对象移除MappedHost对象
 	 * @param host Host对象
 	 */
-	public synchronized void removeHost(Host host) {
-		mapper.remove(host.getName());
+	synchronized void removeHost(Host host) {
+		removeHost(host.getName());
 	}
 	
 	/**
 	 * 根据主机名移除MappedHost对象
 	 * @param hostName
 	 */
-	public synchronized void removeHost(String hostName) {
+	synchronized void removeHost(String hostName) {
+		
+		MappedHost mh = getMappedHost(hostName);
+		if(mh == null)
+			return;
+		
+		hostMapper.remove(mh);
 		mapper.remove(hostName);
 	}
 	
@@ -66,7 +91,10 @@ public final class GlobelMapper {
 	 * @return
 	 */
 	public Host getHost(String hostName) {
-		return getMappedHost(hostName).object;
+		MappedHost mh = getMappedHost(hostName);
+		if(mh == null)
+			return null;
+		return mh.object;
 	}
 	
 	/**
@@ -98,11 +126,53 @@ public final class GlobelMapper {
 			contextName = uri.substring(1, st);
 		}
 		//通过截取后的字符串寻找Context
-		Context context = (Context) host.object.getChildContainer(contextName);
-		if(context != null)
-			return context;
-		//如果没有找到，则转至ROOT目录
-		return (Context) host.object.getChildContainer("ROOT");
+		MappedContext context = getMappedContext(host, contextName);
+		if(context == null)
+			return null;
+		return context.object;
+	}
+	
+	/**
+	 * 向路由器中加入Context映射关系
+	 * @param context Context容器
+	 */
+	void addContext(Context context) {
+		
+		Host host = (Host)context.getParentContainer();
+		MappedHost mappedHost = getMappedHost(host.getName());
+		
+		if(mappedHost == null) {
+			log.error("MappedHost 映射丢失?");
+			return;
+		}
+		
+		Map<String, MappedContext> map = new LinkedHashMap<>();
+		
+		for(Container c : host.getChildContainers()) {
+			MappedContext mappedContext = new MappedContext(c.getName(), context);
+			map.put(c.getName(), mappedContext);
+		}
+		
+		hostMapper.put(mappedHost, map);
+	}
+	
+	/**
+	 * 向路由器中移除Context映射关系
+	 * @param context
+	 */
+	void removeContext(Context context) {
+		Host host = (Host)context.getParentContainer();
+		MappedHost mappedHost = getMappedHost(host.getName());
+		
+		if(mappedHost == null) {
+			return;
+		}
+		
+		hostMapper.remove(mappedHost);
+	}
+	
+	private MappedContext getMappedContext(MappedHost mappedHost, String contextName) {
+		return hostMapper.get(mappedHost).get(contextName);
 	}
 	
 	public Service getService() {
@@ -138,4 +208,21 @@ final class MappedContext extends MapElement<Context> {
 		super(name, object);
 	}
 	
+	@Override
+	public int hashCode() {
+		return name.hashCode();
+	}
+	
+	@Override
+	public boolean equals(Object obj) {
+		
+		if(!(obj instanceof MappedContext)) 
+			return false;
+		
+		MappedContext mc = (MappedContext)obj;
+		
+		if(name.equals(mc.name))
+			return true;
+		return false;
+	}
 }
